+++ pybis/src/python/tests/test_project.py
@@ -18,12 +18,18 @@ def test_create_delete_project(space):
         project.save()
         assert "should not have been created" is None
 
-    project_name = 'project_'+timestamp
+    project_name = 'project_'+timestamp+"_"+str(random.randint(0,1000))
     project=o.new_project(space=space, code=project_name)
     project.save()
 
     project_exists=o.get_project(project_name)
     assert project_exists is not None
+
+    projects_exist=o.get_projects()
+    assert len(projects_exist) > 0
+    first_project = projects_exist[0]
+    assert first_project is not None
+
     project_exists.delete('test project on '+timestamp)
     
     with pytest.raises(ValueError):
@@ -35,7 +41,7 @@ def test_create_project_with_attachment(space):
     o=space.openbis
 
     timestamp = time.strftime('%a_%y%m%d_%H%M%S').upper()
-    project_name = 'project_'+timestamp
+    project_name = 'project_'+timestamp+"_"+str(random.randint(0,1000))
     filename = os.path.join(os.path.dirname(__file__), 'testfile')
 
     if not os.path.exists(filename):