+++ pybis/src/python/CHANGELOG.md
@@ -1,3 +1,10 @@
+## Changes with pybis-1.9.0
+
+* new: search, create, update and delete Property Types
+* new: search, create, update and delete Plugins
+* new: create and update Sample Types
+* freeze features for all entities
+
 ## Changes with pybis-1.8.5
 
 * changed to v3 API when fetching datastores